Move livedata to lifecycle-livedata extension. Add activity, fragment, lifecycle-runtime, material, navigation and recyclerview extensions.
