+++ build.gradle
@@ -10,14 +10,27 @@ buildscript {
         androidCompileSdkVersion = 30
         androidTargerSdkVersion = androidCompileSdkVersion
 
+        libExtensionsActivityVersion = '1.0.0'
         libExtensionsCommonVersion = '1.1.0'
-        libExtensionsLiveDataVersion = '1.1.0'
+        libExtensionsFragmentVersion = '1.0.0'
+        libExtensionsLifecycleLiveDataVersion = '1.1.0'
+        libExtensionsLifecycleRuntimeVersion = '1.0.0'
+        libExtensionsMaterialVersion = '1.0.0'
+        libExtensionsNavigationVersion = '1.0.0'
+        libExtensionsRecyclerViewVersion = '1.0.0'
 
         kotlinVersion = '1.4.21'
         annotationVersion = '1.1.0'
         androidXCoreVersion = '1.3.2'
-        lifecycleLiveDataVersion = '2.3.0-beta01'
+        activityVersion = '1.2.0-rc01'
+        fragmentVersion = '1.3.0-rc01'
+        navigationVersion = '2.3.2'
+        materialVersion = '1.2.1'
+        lifecycleVersion = '2.3.0-rc01'
         liveEventVersion = '1.0.1'
+        recyclerViewVersion = '1.1.0'
+
+        dynamicTextVersion = "3.0.0"
     }
 
     repositories {
