+++ extensions/lifecycle-runtime/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="com.twocoders.extensions.lifecycle.runtime" />
