--- extensions/livedata/build.gradle
@@ -1,33 +1,0 @@
-apply plugin: 'com.android.library'
-apply plugin: 'kotlin-android'
-
-ext.bintrayPublishVersion = libExtensionsLiveDataVersion
-apply from: '../../bintray-publish-config.gradle'
-
-android {
-    compileSdkVersion androidCompileSdkVersion
-
-    defaultConfig {
-        minSdkVersion androidMinSdkVersion
-        targetSdkVersion androidTargerSdkVersion
-        versionName libExtensionsLiveDataVersion
-
-        consumerProguardFiles 'consumer-rules.pro'
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-        }
-    }
-}
-
-dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlinVersion"
-    implementation "androidx.annotation:annotation:$annotationVersion"
-    implementation "androidx.lifecycle:lifecycle-livedata-ktx:$lifecycleLiveDataVersion"
-
-    api "com.github.hadilq.liveevent:liveevent:$liveEventVersion"
-}
