+++ extensions/recyclerview/src/main/java/com/twocoders/extensions/recyclerview/RecyclerViewExtensions.kt
@@ -1,0 +1,18 @@
+package com.twocoders.extensions.recyclerview
+
+import androidx.databinding.BindingAdapter
+import androidx.recyclerview.widget.GridLayoutManager
+import androidx.recyclerview.widget.RecyclerView
+import androidx.recyclerview.widget.StaggeredGridLayoutManager
+import com.twocoders.extensions.common.logd
+
+@BindingAdapter(value = ["spanCount"])
+fun RecyclerView.setSpanCount(spanCount: Int) {
+    with(layoutManager) {
+        when (this) {
+            is StaggeredGridLayoutManager -> this.spanCount = spanCount
+            is GridLayoutManager -> this.spanCount = spanCount
+            else -> logd("SpanCount cannot be assigned to the unsupported RecyclerView LayoutManager: $this")
+        }
+    }
+}
